% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_prior_infections.R
\name{calculate_prior_infections}
\alias{calculate_prior_infections}
\title{A priori probability of further Infections}
\usage{
calculate_prior_infections(
  negative_persons,
  infected_persons,
  event,
  p_one = NULL,
  infect_average = NULL
)
}
\arguments{
\item{negative_persons}{Number of people without the infectious persons.}

\item{infected_persons}{Number of infected people.}

\item{event}{Characters, event type given as characters, currently: "school" or "day_care_center".}

\item{p_one}{Number, this is a placeholder}

\item{infect_average}{Number, this is a placeholder}
}
\value{
The a priori probability y.
}
\description{
Calculates the a priori probability of how many people are infected in one event.
}
\details{
The probability is beta-binomial distributed. The values for p1 and infection_average for
the events "school" and "day_care_center" are from Schoeps et al. (2021).
}
\examples{
calculate_prior_infections(negative_persons = 23,
                           infected_persons = 2,
                           event = "school")

}
\references{
Schoeps A et al. (2021) "Surveillance of SARS-CoV-2 transmission in educational institutions,
August to December 2020, Germany". \emph{Epidemiology and Infection} 149, E213: 1-9.
}
