% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{generate_data_extended}
\alias{generate_data_extended}
\title{Generate data extended}
\usage{
generate_data_extended(
  M = 20,
  d = matrix(data = 1, nrow = 1, ncol = 2),
  S = c(12)
)
}
\arguments{
\item{M}{The size of the group without infected, default is twenty.}

\item{d}{A matrix with the test dates, default is matrix(data = 1, nrow = 1, ncol = 2).}

\item{S}{A vector with the sizes of the subgroups, default is c(12).}
}
\value{
The dataframe.
}
\description{
Creates a dataframe suitable as input for \code{\link{calculate_likelihood_negative_tests}}.
}
