% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_test_sensitivities.R
\name{get_test_sensitivities}
\alias{get_test_sensitivities}
\title{Generate info}
\usage{
get_test_sensitivities(df)
}
\arguments{
\item{df}{Dataframe, this is a placeholder}
}
\value{
The dataframe.
}
\description{
Creates a dataframe with day specific test sensitivity and specificity of
PCR and Antigen tests.
}
\examples{
get_test_sensitivities()
df <- data.frame(
             "PCR" = c(0, 0, 0, 0.04, 0.34, 0.64, 0.76, 0.79, 0.80, 0.79,
                      0.77, 0.74, 0.71, 0.67, 0.62, 0.58, 0.54, 0.49, 0.44,
                      0.40, 0.37, 0.33),
             "Antigen" = c(0, 0, 0, 0.03, 0.13, 0.40,  0.64, 0.69, 0.70, 0.69,
                          0.62, 0.52, 0.40, 0.29, 0.21, 0.17, 0.13, 0.11,
                          0.08, 0.07, 0.05, 0.04)
                )
get_test_sensitivities(df)
}
