% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{p_onePrimaryMore}
\alias{p_onePrimaryMore}
\title{One more primary a priori probability}
\usage{
p_onePrimaryMore(yCurrent, y1)
}
\arguments{
\item{yCurrent}{The current prior distribution.}

\item{y1}{The prior distribution of one single primary case.}
}
\value{
The a priori probability y.
}
\description{
Calculates the a priori probability y for one primary case more
by using the current prior distribution and the prior distribution of
one single primary case.
}
